requirements.txt
pip3 install git+https://github.com/Swind/pure-python-adb.git
pyusb